Notify abort event to mbed-htrun via stdin pipe

+++ mbed_greentea/mbed_test_api.py
@@ -19,7 +19,7 @@ Author: Przemyslaw Wirkus <Przemyslaw.wirkus@arm.com>
 
 import re
 import sys
-from time import time
+from time import time, sleep
 from Queue import Queue, Empty
 from threading import Thread
 from subprocess import call, Popen, PIPE
@@ -174,11 +174,28 @@ def run_host_test(image_path,
 
         def stop(self):
             self.active = False
+
+            # Try stopping mbed-host-test
             try:
-                self.proc.terminate()
-            except Exception as e:
-                print "ProcessObserver.stop(): %s" % str(e)
+                self.proc.stdin.close()
+            finally:
                 pass
+
+            # Give 5 sec for mbedhtrun to exit
+            ret_code = None
+            for i in range(5):
+                ret_code = self.proc.poll()
+                # A None value indicates that the process hasn't terminated yet.
+                if ret_code is not None:
+                    break
+                sleep(1)
+
+            if ret_code is None:            # Kill it
+                print 'Terminating mbed-host-test(mbedhtrun) process (PID %s)' % self.proc.pid
+                try:
+                    self.proc.terminate()
+                except Exception as e:
+                    print "ProcessObserver.stop(): %s" % str(e)
 
     def get_char_from_queue(obs):
         """ Get character from queue safe way
@@ -256,7 +273,7 @@ def run_host_test(image_path,
         if verbose:
             gt_logger.gt_log_tab("calling mbedhtrun: %s"% " ".join(cmd))
             gt_logger.gt_log("mbed-host-test-runner: started")
-        proc = Popen(cmd, stdout=PIPE)
+        proc = Popen(cmd, stdin=PIPE, stdout=PIPE)
         obs = ProcessObserver(proc)
 
     result = None
